:fire: hoist engine out of base UI
