+++ src/ui.test.ts
@@ -13,7 +13,7 @@ function parseAndReturnFirstSpritePixels(code: string) {
     engine.setLevel(0)
     const cell = engine.getCurrentLevelCells()[0][0]
     // console.log(cell.getSprites())
-    UI.setGameEngine(engine)
+    UI.setGameData(engine.getGameData())
     const pixels = UI.getPixelsForCell(cell)
     UI.destroy()
     return { pixels, data }
